chore: move EventBanner to banners
