+++ website-frontend/src/lib/components/banners/EventBanner.svelte
